requirements.txt
flask==3.0.0
pymongo==4.6.1
python-dotenv==1.0.0
reportlab==4.0.9
pillow==11.0.0
gunicorn==21.2.0
waitress==2.1.2
